refactor(Compatibility) rename spec -> specification
