+++ lib/graphql/compatibility.rb
@@ -1,3 +1,3 @@
-require "graphql/compatibility/execution_spec"
-require "graphql/compatibility/query_parser_spec"
-require "graphql/compatibility/schema_parser_spec"
+require "graphql/compatibility/execution_specification"
+require "graphql/compatibility/query_parser_specification"
+require "graphql/compatibility/schema_parser_specification"
