--- spec/graphql/compatibility/execution_spec_spec.rb
@@ -1,3 +1,0 @@
-require "spec_helper"
-
-SerialExecutionSuite = GraphQL::Compatibility::ExecutionSpec.build_suite(GraphQL::Query::SerialExecution)
