+++ spec/graphql/compatibility/execution_specification_spec.rb
@@ -1,0 +1,3 @@
+require "spec_helper"
+
+SerialExecutionSuite = GraphQL::Compatibility::ExecutionSpecification.build_suite(GraphQL::Query::SerialExecution)
