--- spec/graphql/compatibility/query_parser_spec_spec.rb
@@ -1,5 +1,0 @@
-require "spec_helper"
-
-BuiltInParserSuite = GraphQL::Compatibility::QueryParserSpec.build_suite do |query_string|
-  GraphQL::Language::Parser.parse(query_string)
-end
