+++ spec/graphql/compatibility/schema_parser_specification_spec.rb
@@ -1,0 +1,5 @@
+require "spec_helper"
+
+BuiltInParserSuite = GraphQL::Compatibility::SchemaParserSpecification.build_suite do |query_string|
+  GraphQL::Language::Parser.parse(query_string)
+end
